キャスト関係の警告をわかるやつだけ減らした

+++ WinterGame_2025/Source/Game.h
@@ -28,8 +28,8 @@ namespace GlobalConstants
 	// マップチップ情報
 	constexpr int kChipSize = 16;
 	constexpr float kDrawScale = 3.0f;
-	constexpr int kDrawChipSize = kChipSize * kDrawScale;
-	constexpr int kDrawChipSizeHalf = kChipSize * kDrawScale / 2;
+	constexpr int kDrawChipSize = static_cast<int>(kChipSize * kDrawScale);
+	constexpr int kDrawChipSizeHalf = static_cast<int>(kChipSize * kDrawScale / 2);
 
 	// 共通フォント名
 	constexpr char kMainFontName[] = "廻想体 ネクスト UP B";
