+++ WinterGame_2025/Source/GameObjects/Enemies/WalkBoss.cpp
@@ -305,8 +305,8 @@ void WalkBoss::Draw(Vector2 offset)
 			if (_frame % kDeathFrickerFrame == 0)
 			{
 				Vector2 explosionPosGap;
-				explosionPosGap.x = GetRand(kDeathExplosionGapDis * 2) - kDeathExplosionGapDis;
-				explosionPosGap.y = GetRand(kDeathExplosionGapDis * 2) - kDeathExplosionGapDis;
+				explosionPosGap.x = static_cast<float>(GetRand(kDeathExplosionGapDis * 2) - kDeathExplosionGapDis);
+				explosionPosGap.y = static_cast<float>(GetRand(kDeathExplosionGapDis * 2) - kDeathExplosionGapDis);
 				_pEffectManager->Create(GetColliderPos() + explosionPosGap, EffectType::ExplosionSmall);
 			}
 		}
@@ -330,7 +330,7 @@ void WalkBoss::Draw(Vector2 offset)
 			_drawBarrierFrame--;
 		}
 		float progress = static_cast<float>(_drawBarrierFrame) / kDrawBarrierMaxFrame;
-		SetDrawBlendMode(DX_BLENDMODE_ALPHA, progress * 128);
+		SetDrawBlendMode(DX_BLENDMODE_ALPHA, static_cast<int>(progress * 128));
 		_barrierAnim.Draw(GetColliderPos() - offset, false);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 	}
