+++ WinterGame_2025/Source/Systems/Animation.cpp
@@ -71,9 +71,9 @@ void Animation::Draw(Vector2 pos, bool isTurn)
 	// 現在のコマ数が最大コマ数なら描画しない
 	if (_nowAnimNum != _maxAnimNum)
 	{
-		DrawRectRotaGraph(pos.x + _drawOffset.x, pos.y + _drawOffset.y,
-			_frameSize.x * _nowAnimNum, _animIndex * _frameSize.y,
-			_frameSize.x, _frameSize.y,
+		DrawRectRotaGraph(static_cast<int>(pos.x + _drawOffset.x), static_cast<int>(pos.y + _drawOffset.y),
+			static_cast<int>(_frameSize.x) * _nowAnimNum, _animIndex * static_cast<int>(_frameSize.y),
+			static_cast<int>(_frameSize.x), static_cast<int>(_frameSize.y),
 			_scale, _rotate, _drawHandle, true, isTurn);
 	}
 }
@@ -82,9 +82,9 @@ void Animation::Draw(int drawHandle, Vector2 pos, bool isTurn)
 {
 	if (_nowAnimNum != _maxAnimNum)
 	{
-		DrawRectRotaGraph(pos.x, pos.y,
-			_frameSize.x * _nowAnimNum, _animIndex * _frameSize.y,
-			_frameSize.x, _frameSize.y,
+		DrawRectRotaGraph(static_cast<int>(pos.x), static_cast<int>(pos.y),
+			static_cast<int>(_frameSize.x) * _nowAnimNum, _animIndex * static_cast<int>(_frameSize.y),
+			static_cast<int>(_frameSize.x), static_cast<int>(_frameSize.y),
 			_scale, _rotate, drawHandle, true, isTurn);
 	}
 }
