+++ WinterGame_2025/Source/UI/BossHPUI.cpp
@@ -19,13 +19,13 @@ namespace
 	constexpr float kDrawScale = 0.4f;
 	constexpr int kDrawPosX = 1400;
 	constexpr int kDrawPosY = 130;
-	constexpr int kDrawPosL = kDrawPosX - (kGraphSizeW * kDrawScale) / 2;
-	constexpr int kDrawPosT = kDrawPosY - (kGraphSizeH * kDrawScale) / 2;
+	constexpr int kDrawPosL = static_cast<int>(kDrawPosX - (kGraphSizeW * kDrawScale) / 2);
+	constexpr int kDrawPosT = static_cast<int>(kDrawPosY - (kGraphSizeH * kDrawScale) / 2);
 	// バーの描画情報
-	constexpr int kDrawBarL = kDrawPosL + (kBarL * kDrawScale);
-	constexpr int kDrawBarR = kDrawPosL + (kBarR * kDrawScale);
-	constexpr int kDrawBarT = kDrawPosT + (kBarT * kDrawScale);
-	constexpr int kDrawBarB = kDrawPosT + (kBarB * kDrawScale) + 1;	// 隙間を埋めるために1足す
+	constexpr int kDrawBarL = static_cast<int>(kDrawPosL + (kBarL * kDrawScale));
+	constexpr int kDrawBarR = static_cast<int>(kDrawPosL + (kBarR * kDrawScale));
+	constexpr int kDrawBarT = static_cast<int>(kDrawPosT + (kBarT * kDrawScale));
+	constexpr int kDrawBarB = static_cast<int>(kDrawPosT + (kBarB * kDrawScale)) + 1;	// 隙間を埋めるために1足す
 	constexpr int kBarLength = kDrawBarR - kDrawBarL;
 
 	// 透明になる領域を余分に確保する量
@@ -54,9 +54,9 @@ void BossHPUI::Init()
 void BossHPUI::Update()
 {
 	// ボスのhpからバーの長さを出す
-	_barLength = static_cast<float>(_pEnemyManager->GetTotalBossHp()) / static_cast<float>(_bossMaxHp) * kBarLength;
+	_barLength = static_cast<int>(static_cast<float>(_pEnemyManager->GetTotalBossHp()) / static_cast<float>(_bossMaxHp) * kBarLength);
 	// lerpでいい感じに減らす
-	_drawBarLength = std::lerp(_drawBarLength, _barLength, 0.05f);
+	_drawBarLength = static_cast<int>(std::lerp(_drawBarLength, _barLength, 0.05f));
 }
 
 void BossHPUI::Draw()
@@ -79,11 +79,11 @@ void BossHPUI::Draw()
 	// 透明度をlerpでいい感じに変える
 	if (isPlayerNear || isEnemyNear)
 	{
-		_alpha = std::lerp(_alpha, 64, 0.2f);
+		_alpha = static_cast<int>(std::lerp(_alpha, 64, 0.2f));
 	}
 	else
 	{
-		_alpha = std::lerp(_alpha, 255, 0.2f);
+		_alpha = static_cast<int>(std::lerp(_alpha, 255, 0.2f));
 	}
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, _alpha);
 	DrawBox(kDrawBarL + _drawBarLength, kDrawBarT, kDrawBarR, kDrawBarB, 0x000000, true);	// HPないとこの黒
